requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Azure AI Projects (Foundry SDK)
azure-ai-projects==2.0.0b4
azure-identity>=1.19.0
aiohttp>=3.11.0

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.1

# WebSocket support (included with uvicorn[standard])
websockets==14.1

# Environment variables
python-dotenv==1.0.1

# CORS
starlette==0.41.3

# Utilities
httpx==0.28.1
tenacity==9.0.0

# Database
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
alembic==1.18.4

# Tracing / Observability
opentelemetry-sdk==1.36.0
opentelemetry-api==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-exporter-otlp==1.36.0
azure-core-tracing-opentelemetry>=1.0.0b12
azure-monitor-opentelemetry==1.8.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
